requirements.txt
streamlit
mysql-connector-python
streamlit-option-menu
seaborn
matplotlib